Create new esbuild workflow
