+++ aws_lambda_builders/workflows/__init__.py
@@ -11,3 +11,4 @@ import aws_lambda_builders.workflows.java_gradle
 import aws_lambda_builders.workflows.java_maven
 import aws_lambda_builders.workflows.dotnet_clipackage
 import aws_lambda_builders.workflows.custom_make
+import aws_lambda_builders.workflows.nodejs_npm_esbuild
