+++ aws_lambda_builders/workflows/nodejs_npm/utils.py
@@ -9,8 +9,6 @@ import subprocess
 import shutil
 import json
 
-EXPERIMENTAL_FLAG_ESBUILD = "experimentalEsbuild"
-
 
 class OSUtils(object):
 
@@ -55,10 +53,3 @@ class OSUtils(object):
     def parse_json(self, path):
         with open(path) as json_file:
             return json.load(json_file)
-
-
-def is_experimental_esbuild_scope(experimental_flags):
-    """
-    A function which will determine if experimental esbuild scope is active
-    """
-    return bool(experimental_flags) and EXPERIMENTAL_FLAG_ESBUILD in experimental_flags
