+++ aws_lambda_builders/workflows/nodejs_npm_esbuild/__init__.py
@@ -1,0 +1,5 @@
+"""
+Builds NodeJS Lambda functions using NPM dependency manager with esbuild bundler
+"""
+
+from .workflow import NodejsNpmEsbuildWorkflow
