+++ tests/integration/workflows/nodejs_npm/test_nodejs_npm.py
@@ -135,21 +135,6 @@ class TestNodejsNpmWorkflow(TestCase):
 
         self.assertIn("No matching version found for minimal-request-promise@0.0.0-NON_EXISTENT", str(ctx.exception))
 
-    def test_fails_if_package_json_is_broken(self):
-
-        source_dir = os.path.join(self.TEST_DATA_FOLDER, "broken-package")
-
-        with self.assertRaises(WorkflowFailedError) as ctx:
-            self.builder.build(
-                source_dir,
-                self.artifacts_dir,
-                self.scratch_dir,
-                os.path.join(source_dir, "package.json"),
-                runtime=self.runtime,
-            )
-
-        self.assertIn("NodejsNpmBuilder:ParseManifest", str(ctx.exception))
-
     def test_builds_project_with_remote_dependencies_without_download_dependencies_with_dependencies_dir(self):
         source_dir = os.path.join(self.TEST_DATA_FOLDER, "npm-deps")
