+++ tests/integration/workflows/nodejs_npm_esbuild/testdata/broken-package/package.json
