+++ tests/integration/workflows/nodejs_npm_esbuild/testdata/esbuild-binary/package.json
