+++ tests/integration/workflows/nodejs_npm_esbuild/testdata/no-deps-esbuild/package.json
