+++ tests/integration/workflows/nodejs_npm_esbuild/testdata/with-deps-esbuild-typescript/package.json
