+++ tests/integration/workflows/nodejs_npm_esbuild/testdata/with-deps-esbuild/excluded.js
