+++ tests/unit/workflows/nodejs_npm/test_workflow.py
@@ -1,21 +1,16 @@
 from unittest import TestCase
-from mock import patch, call
+from mock import patch
 
-from aws_lambda_builders.exceptions import WorkflowFailedError
 from aws_lambda_builders.actions import CopySourceAction, CleanUpAction, CopyDependenciesAction, MoveDependenciesAction
 from aws_lambda_builders.architecture import ARM64
 from aws_lambda_builders.workflows.nodejs_npm.workflow import NodejsNpmWorkflow
-from aws_lambda_builders.workflows.nodejs_npm.esbuild import SubprocessEsbuild
 from aws_lambda_builders.workflows.nodejs_npm.actions import (
     NodejsNpmPackAction,
     NodejsNpmInstallAction,
     NodejsNpmrcCopyAction,
     NodejsNpmrcCleanUpAction,
     NodejsNpmLockFileCleanUpAction,
-    NodejsNpmCIAction,
-    EsbuildBundleAction,
 )
-from aws_lambda_builders.workflows.nodejs_npm.utils import EXPERIMENTAL_FLAG_ESBUILD
 
 
 class FakePopen:
@@ -164,135 +159,6 @@ class TestNodejsNpmWorkflow(TestCase):
         self.assertIsInstance(workflow.actions[6], NodejsNpmrcCleanUpAction)
         self.assertIsInstance(workflow.actions[7], NodejsNpmLockFileCleanUpAction)
         self.assertIsInstance(workflow.actions[8], NodejsNpmLockFileCleanUpAction)
-
-    def test_workflow_sets_up_npm_actions_with_bundler_if_manifest_requests_it(self):
-
-        self.osutils.parse_json.side_effect = [{"aws_sam": {"bundler": "esbuild"}}]
-
-        self.osutils.file_exists.side_effect = [True, False, False]
-
-        workflow = NodejsNpmWorkflow(
-            "source",
-            "artifacts",
-            "scratch_dir",
-            "manifest",
-            osutils=self.osutils,
-            experimental_flags=[EXPERIMENTAL_FLAG_ESBUILD],
-        )
-
-        self.assertEqual(len(workflow.actions), 2)
-
-        self.assertIsInstance(workflow.actions[0], NodejsNpmInstallAction)
-
-        self.assertIsInstance(workflow.actions[1], EsbuildBundleAction)
-
-        self.osutils.parse_json.assert_called_with("manifest")
-
-        self.osutils.file_exists.assert_has_calls(
-            [call("source/package-lock.json"), call("source/npm-shrinkwrap.json")]
-        )
-
-    def test_workflow_fails_if_manifest_parsing_fails(self):
-
-        self.osutils.parse_json.side_effect = OSError("boom!")
-
-        with self.assertRaises(WorkflowFailedError) as raised:
-            NodejsNpmWorkflow("source", "artifacts", "scratch_dir", "manifest", osutils=self.osutils)
-
-        self.assertEqual(raised.exception.args[0], "NodejsNpmBuilder:ParseManifest - boom!")
-
-        self.osutils.parse_json.assert_called_with("manifest")
-
-    def test_sets_up_esbuild_search_path_from_npm_bin(self):
-
-        self.popen.out = b"project/bin"
-        self.osutils.parse_json.side_effect = [{"aws_sam": {"bundler": "esbuild"}}]
-
-        workflow = NodejsNpmWorkflow(
-            "source",
-            "artifacts",
-            "scratch_dir",
-            "manifest",
-            osutils=self.osutils,
-            experimental_flags=[EXPERIMENTAL_FLAG_ESBUILD],
-        )
-
-        self.osutils.popen.assert_called_with(["npm", "bin"], stdout="PIPE", stderr="PIPE", cwd="source")
-
-        esbuild = workflow.actions[1].subprocess_esbuild
-
-        self.assertIsInstance(esbuild, SubprocessEsbuild)
-
-        self.assertEqual(esbuild.executable_search_paths, ["project/bin"])
-
-    def test_sets_up_esbuild_search_path_with_workflow_executable_search_paths_after_npm_bin(self):
-
-        self.popen.out = b"project/bin"
-        self.osutils.parse_json.side_effect = [{"aws_sam": {"bundler": "esbuild"}}]
-
-        workflow = NodejsNpmWorkflow(
-            "source",
-            "artifacts",
-            "scratch_dir",
-            "manifest",
-            osutils=self.osutils,
-            executable_search_paths=["other/bin"],
-            experimental_flags=[EXPERIMENTAL_FLAG_ESBUILD],
-        )
-
-        self.osutils.popen.assert_called_with(["npm", "bin"], stdout="PIPE", stderr="PIPE", cwd="source")
-
-        esbuild = workflow.actions[1].subprocess_esbuild
-
-        self.assertIsInstance(esbuild, SubprocessEsbuild)
-
-        self.assertEqual(esbuild.executable_search_paths, ["project/bin", "other/bin"])
-
-    def test_workflow_uses_npm_ci_if_lockfile_exists(self):
-
-        self.osutils.parse_json.side_effect = [{"aws_sam": {"bundler": "esbuild"}}]
-        self.osutils.file_exists.side_effect = [True, True]
-
-        workflow = NodejsNpmWorkflow(
-            "source",
-            "artifacts",
-            "scratch_dir",
-            "manifest",
-            osutils=self.osutils,
-            experimental_flags=[EXPERIMENTAL_FLAG_ESBUILD],
-        )
-
-        self.assertEqual(len(workflow.actions), 2)
-
-        self.assertIsInstance(workflow.actions[0], NodejsNpmCIAction)
-
-        self.assertIsInstance(workflow.actions[1], EsbuildBundleAction)
-
-        self.osutils.file_exists.assert_has_calls([call("source/package-lock.json")])
-
-    def test_workflow_uses_npm_ci_if_shrinkwrap_exists(self):
-
-        self.osutils.parse_json.side_effect = [{"aws_sam": {"bundler": "esbuild"}}]
-        self.osutils.file_exists.side_effect = [True, False, True]
-
-        workflow = NodejsNpmWorkflow(
-            "source",
-            "artifacts",
-            "scratch_dir",
-            "manifest",
-            osutils=self.osutils,
-            experimental_flags=[EXPERIMENTAL_FLAG_ESBUILD],
-        )
-
-        self.assertEqual(len(workflow.actions), 2)
-
-        self.assertIsInstance(workflow.actions[0], NodejsNpmCIAction)
-
-        self.assertIsInstance(workflow.actions[1], EsbuildBundleAction)
-
-        self.osutils.file_exists.assert_has_calls(
-            [call("source/package-lock.json"), call("source/npm-shrinkwrap.json")]
-        )
 
     def test_must_validate_architecture(self):
         self.osutils.is_windows.side_effect = [False, False]
